splits the bam2cram an SV annotation workflows for long-read data
