+++ configs/defaults/seqr_loader_long_read.toml
@@ -3,14 +3,3 @@ name = 'seqr_loader_long_read'
 dataset_gcp_project = 'seqr-308602'
 status_reporter = 'metamist'
 bam_to_cram_analysis_type = 'pacbio_cram'
-
-[references.gatk_sv]
-# a couple of annotation arguments are not files
-# github.com/broadinstitute/gatk-sv/blob/main/inputs/templates/test/AnnotateVcf/AnnotateVcf.json.tmpl#L4-L8
-external_af_population = ['ALL', 'AFR', 'AMR', 'EAS', 'EUR']
-external_af_ref_bed_prefix = 'gnomad_v2.1_sv'
-
-[images]
-gatk_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv/gatk:2023-07-13-4.4.0.0-43-gd79823f9c-NIGHTLY-SNAPSHOT"
-sv_base_mini_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv/sv-base-mini:5994670"
-sv_pipeline_docker = "australia-southeast1-docker.pkg.dev/cpg-common/images/sv/sv-pipeline:2023-09-13-v0.28.3-beta-af8362e3"
