+++ main.py
@@ -30,6 +30,7 @@ from cpg_workflows.stages.happy_validation import (
     ValidationParseHappy,
 )
 from cpg_workflows.stages.large_cohort import AncestryPlots, Frequencies, LoadVqsr
+from cpg_workflows.stages.seqr_loader_long_read.long_read_sv_annotation import AnnotateLongReadSVs, ReFormatPacBioSVs
 from cpg_workflows.stages.mito import MitoReport
 from cpg_workflows.stages.outrider import Outrider
 from cpg_workflows.stages.seqr_loader import AnnotateDataset, DatasetVCF, MtToEs
@@ -41,6 +42,10 @@ from cpg_workflows.workflow import StageDecorator, run_workflow
 WORKFLOWS: dict[str, list[StageDecorator]] = {
     'talos': [ValidateMOI, CreateTalosHTML, GenerateSeqrFile],
     'exomiser': [RunExomiser, ExomiserSeqrTSV],
+    'long_read_sv_annotation': [
+        ReFormatPacBioSVs,
+        AnnotateLongReadSVs,
+    ],
     'pre_alignment': [FastQCMultiQC],
     'seqr_loader': [
         DatasetVCF,
